pyproject.toml
[project]
name = "nextgen-mcp"
version = "0.1.0"
description = "Production-grade MCP boilerplate"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.1",
    "python-dotenv>=1.2.2",
]

[project.scripts]
nextgen-mcp = "nextgen_mcp.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nextgen_mcp"]
